web_research_agent/requirements.txt
openai
tavily-search
python-dotenv